requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.2.0
pdf2docx==0.5.8
python-docx==1.1.0
moviepy==1.0.3
aiofiles==23.2.1
